==> notebooks/2_Cleanup_Location.ipynb <==
import pandas as pd
import pickle
import nltk
import re
import string
import numpy as np

from pprint import pprint


with open('xl.pkl', 'rb') as picklefile:
    df = pickle.load(picklefile)

### Clean up text of tweets and identify retweets versus original 

#### keep puncuation for Vader sentiment analysis

df.head(2)

def remove_retweet_url(tweet):
    """remove RT and links in tweet""" 
    tweet.strip()
    tweet = re.sub(r"(https?\://)\S+", "", tweet)
    if tweet.startswith('RT'):
        tweet = re.sub(r"(RT\s@.*:)", "", tweet)
    return tweet

df['clean'] = df.text.apply(lambda x : remove_retweet_url(x))

def countRetweet(row):
    """ return 1 if retweet, 0 if original"""
    words = row['text'].split(' ')
    if words[0] == 'RT':
        return 1
    else:
        return 0

def countOriginal(row):
    """ return 0 if retweet, 1 if original"""
    words = row['text'].split(' ')
    if words[0] != 'RT':
        return 1
    else:
        return 0

df['retweet'] = df.apply(countRetweet, axis = 1)

df['original_tweet'] = df.apply(countOriginal, axis = 1)

len(df)

### Location Retrieval

#### Stripping puncuation for location analysis. Build a dictionary of lat/long cooridinates for the top 2500 locations. The google geocode API accepts 2500 calls per day. 

def lower_alpha_num(corpus):
    # convert to lower case
    corpus = map(str.lower, corpus)
    
    # remove alpha-numerical words
    corpus = map(lambda x: re.sub(r"""\w*\d\w*""", '', x), corpus)
    corpus = list(corpus)
    s = ''.join(corpus)
    return s

def remove_punct(corpus):
    # regular expression to remove punctuation
    punc_re = re.compile('[%s]' % re.escape(string.punctuation))

    corpus = map(lambda x: punc_re.sub('', x), corpus)
    corpus = list(corpus)
    s = ''.join(corpus)
    return s

# Select the rows of df where Location is not NaN
from copy import deepcopy
df_location = df[df['location'].notnull()]
df_location = deepcopy(df_location)

df_location['location'] = df_location.location.apply(lambda x: lower_alpha_num(x))

df_location['location'] = df_location.location.apply(lambda x: remove_punct(x))

df_location.head(5)

#### Get a list of the top 2480 locations

# locations = [i for i in df_location_final.location.value_counts()[:2480].index]
# locations

### Call the google Geocode API

import urllib
import json
from urllib.parse import urlparse
import urllib.parse
import urllib.request
import time

import geocoder
import time

googleGeocodeUrl = 'http://maps.googleapis.com/maps/api/geocode/json?'

def get_coordinates(locations, from_sensor=False):
    ''' return a dictionary of lat, long and region'''
    location_dict = {}
    for query in locations:
        query = query.encode('utf-8')
        params = {
            'address': query,
            'sensor': "true" if from_sensor else "false"
        }
        
        url = googleGeocodeUrl + urllib.parse.urlencode(params)
        try:
            json_response = urllib.request.urlopen(url)
            response = json.loads(json_response.read())
            if response['results']:
                location = response['results'][0]['geometry']['location']
                latitude, longitude = location['lat'], location['lng']
                region = response['results'][0]['address_components'][0]['long_name']
                location_dict[query] = latitude, longitude, region
                time.sleep(1)
            else:
                try:
                    location_dict[query] == None
                    time.sleep(1)
                except KeyError:
                    pass
        
        except urllib.error.HTTPError as err:
            if err == 400: 
                location_dict[query] == None
            
    return location_dict
    

with open('dict.pkl', 'rb') as picklefile:
    d = pickle.load(picklefile)

def convert(data):
    if isinstance(data, bytes):  return data.decode('utf-8')
    if isinstance(data, dict):   return dict(map(convert, data.items()))
    if isinstance(data, tuple):  return tuple(map(convert, data))
    return data

d = convert(d)

#### After pulling the lat/long from the Google Maps API, I used Python geocoder wrapper to build a dictionary mapping lat/long to city, state, country, county

def reverse_geo(d):
    '''provided lat/long, returns a dictionary of city, state, country, county and postal formatted state'''
    reverse_dict = {}
    for k, v in list(d.items()):
        g = geocoder.google([d[k][0], d[k][1]], method='reverse')
        reverse_dict[k] = g.lat, g.lng, g.city, g.state_long, g.country_long, g.county, g.state
        time.sleep(1)
    return reverse_dict

# reverse_dict = reverse_geo(d)

with open('reverse_dict.pkl', 'rb') as picklefile:
    reverse_dict = pickle.load(picklefile)

list(reverse_dict.items())[0:4]

# with open('df_location_final.pkl', 'rb') as picklefile:
#     df_location_final = pickle.load(picklefile)

location_keys = ['latitude','longitude','city','state', 'country', 'county', 'postal_state']


def get_location(row, reverse_dict, index):
    try:
        if reverse_dict[row][index] !=None:
            loc = reverse_dict[row][index]
        else:
            loc = ''
        return loc
    except KeyError:
        pass

#### Add columns to dataframe with city, state, county etc...

for i, item in enumerate(location_keys):
    df_location[item] = df_location.location.apply(lambda x: get_location(x, reverse_dict,i))

df_location.head(5) 

# with open('location_df.pkl', 'wb') as picklefile:
#     pickle.dump(df_location, picklefile)

# with open('df_location_final.pkl', 'rb') as picklefile:
#     df_location_final = pickle.load(picklefile)

df_location.head(5)

### Group data by artist, location and tweets

grouped = df_location.groupby(['artist', 'latitude', 'longitude','city', 'state', 'country']).count()


grouped = grouped.reset_index()[['artist', 'latitude', 'longitude', 'city', 'state', 'country', 'id']].sort_values(by='id', ascending=False)

grouped = grouped.rename(columns = {'id':'tweets'}).reset_index(drop=True)

grouped.head(10)

#### During checks on data, I discovered that Independence Kansas was returned from users who only specified "United States". Let's remove it. 

grouped_test = grouped[grouped['city'] != 'Independence']

#### write to csv for plotting

#grouped_test.to_csv('location_world_backup.csv', index=False)
