==> notebooks/3_Sentiments.ipynb <==
### Sentiment analysis using Vader: https://github.com/cjhutto/vaderSentiment

import pandas as pd
import pickle
import nltk

# nltk_path = '/home/ubuntu/nltk_data'
# nltk.data.path.insert(0, nltk_path)

with open('location_df.pkl', 'rb') as picklefile:
    df_location_final = pickle.load(picklefile)

import vaderSentiment
from vaderSentiment.vaderSentiment import SentimentIntensityAnalyzer

analyser = SentimentIntensityAnalyzer()

def print_sentiment_scores(tweet):
    snt = analyser.polarity_scores(tweet)
    return snt

df_location_final['sentiments'] = df_location_final.clean.apply(lambda x: print_sentiment_scores(x))

def get_sentiment(row,sent):
    sentiment = row['sentiments'][sent]
    return sentiment

sentiment_keys = ['compound','neg','neu','pos']
for i in sentiment_keys:
    df_location_final[i] = df_location_final.apply(lambda x: get_sentiment(x,i), axis=1)

df_location_final.head(3)
