==> notebooks/4_World_Plotly.ipynb <==
### Map all tweets using Plotly

import pickle
import numpy as np
import plotly
import pandas as pd

df = pd.read_csv('location_world_backup.csv')

df_world = df

sorter = {'Adele': 0, 'Radiohead': 1,'Thom Yorke': 2, 'Sigur Ros': 3, 'Queens of the Stone Age': 4, 'Kingkrule': 5, 'Kay Tranada': 6, 'Rostam': 7, 'Ibeyi': 8}

df_world['rank'] = df_world['artist'].map(sorter)

df_world = df_world.sort_values(by='rank').reset_index(drop=True)

plotly.tools.set_credentials_file(username='brianturn', api_key='YCuD2uwOkDlx7ib1Zud9')

import plotly.plotly as py

df_world['text'] = df_world['artist'] + '<br>' + (df_world['tweets']).astype(str) + ' mentions'
limits = [(0,1237),(1237,2374),(2374,3176),(3176,3603),(3603,3973),(3973,4214)]
colors = ['rgb(255,0,0)','rgb(0,0,255)','rgb(0,255,0)','rgb(128,0,255)','rgb(255,128,0)','rgb(255,255,0)']
cities = []
scale = 5

for i in range(len(limits)):
    lim = limits[i]
    df_sub = df_world[lim[0]:lim[1]]
    city = dict(
        type = 'scattergeo',
        locationmode = 'usa',
        lon = df_sub['longitude'],
        lat = df_sub['latitude'],
        text = df_sub['text'],
        marker = dict(
            size = df_sub['tweets']/scale,
            color = colors[i],
            line = dict(width=0.5, color='rgb(40,40,40)'),
            sizemode = 'area'
        ),
        name = '{0} - {1}'.format(lim[0],lim[1]) )
    cities.append(city)

layout = dict(
        title = 'Twitter Mentions - Adele and Radiohead dominate<br>June, 2017',
        showlegend = True,
        geo = dict(
            projection=dict( type='Albers USA' ),
            font=dict(family='Helvetica', size=18, color='rgb(40,40,40)'),
            showland = True,
            showcoastlines = True,
            landcolor = 'rgb(217, 217, 217)',
            subunitwidth=.5,
            countrywidth=.5,
            subunitcolor="rgb(40,40,40)",
            countrycolor="rgb(255, 255, 255)"
        ),
    )

fig = dict( data=cities, layout=layout )
py.iplot( fig, validate=False, filename='d3-world-tweets' )
